Refactored code, added styling
- Refactored the ORM methods out of the routes folder into their own folder
- Created an api routes folder for if we have time to add API method calls for the ORM methods (extremely low priority)
- Broke the different models into their own files
- Added styling to make it look more like the Figma mockups
