+++ website/__init__.py
@@ -42,23 +42,24 @@ def create_app():
     from .routes.ui.auth import auth
     from .routes.ui.people import people
     from .routes.ui.events import events
-    from .routes.orm.user import user_orm
-    from .routes.orm.event import event_orm
+    from .routes.api.user import user_api
+    from .routes.api.event import event_api
 
     app.register_blueprint(views, url_prefix='/')
     app.register_blueprint(auth, url_prefix='/')
     app.register_blueprint(people, url_prefix='/')
     app.register_blueprint(events, url_prefix='/')
-    app.register_blueprint(user_orm, url_prefix='/')
-    app.register_blueprint(event_orm, url_prefix='/')
+    app.register_blueprint(user_api, url_prefix='/api/')
+    app.register_blueprint(event_api, url_prefix='/api/')
 
-    from .models import User
     
     with app.app_context():
         db.create_all()
 
     login_manager.init_app(app)
 
+    from .models.user import User
+    
     @login_manager.user_loader
     def load_user(id):
         return User.query.get(int(id))
