--- website/models/__init__.py
@@ -1,2 +1,0 @@
-from .event import *
-from .user import *
