+++ website/models/event/__init__.py
@@ -1,0 +1,3 @@
+from .event import Event, EventParticipant
+from .event_occurrence import EventOccurrence
+from .event_tag import EventTag, EventTagRelationship
