+++ website/models/user/__init__.py
@@ -1,0 +1,2 @@
+from .user import User
+from .user_tag import UserTag, UserTagRelationship
