+++ website/routes/api/event.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+event_api = Blueprint('event_route', __name__)
+
+ #TODO: Implement these API methods
