+++ website/routes/ui/auth.py
@@ -1,8 +1,8 @@
 from flask import Blueprint, render_template, request, flash, redirect, url_for
-from ...models import User
+from ...models.user import User
 from flask_login import login_user, login_required, logout_user, current_user
 from ...forms.auth import SignUpForm, LogInForm
-from ..orm.user import create_user
+from ...orm.user.user import create_user
 
 auth = Blueprint('auth', __name__)
