+++ website/routes/ui/events.py
@@ -1,6 +1,6 @@
 from flask import Blueprint, flash, redirect, render_template, request, url_for
 from flask_login import current_user, login_required
-from ..orm.event import read_event, create_event
+from ...orm.event.event import read_event, create_event
 from ...forms.events import CreateEventForm
 
 events = Blueprint('events', __name__)
@@ -76,9 +76,10 @@ def event_create():
 
     return render_template("create-event.html", user=current_user, event_form=event_form)
 
+#TODO: rework this to use the orm methods
 @events.route('/events/<int:event_id>', methods=['GET'])
 def event_details(event_id: int):
     from ... import db
-    from ...models import Event
+    from ...models.event import Event
     event: Event = db.session.query(Event).get(event_id)
     return render_template("event-details.html", user=current_user, event=event)
