+++ website/templates/create-event.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Create Event{% endblock %}
 {% block content %}
-<div class="page-header">
+<div class="page-header mb-3">
     <h1>Create Event</h1>
 </div>
 <form method="POST">
