+++ website/templates/events.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Events{% endblock %} 
 {% block content %}
-  <div class="page-header">
+  <div class="page-header mb-3">
     <h1>Events</h1>
   </div>
   <table class="table">
