+++ website/templates/home.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}Home{% endblock %} 
 {% block content %}
-  <div class="page-header">
+  <div class="page-header mb-3">
     <h1>Home</h1>
   </div>
 {% endblock %}
