+++ website/templates/login.html
@@ -1,18 +1,27 @@
 {% extends "base.html" %} {% block title %}Login{% endblock %} {% block content
 %}
 <form method="POST">
-  <div class="page-header">
-    <h1>Log In</h1>
-  </div>
-  <div class="form-group">
-    {{ form.email.label() }}
-    {{ form.email() }}
-  </div>
-  <div class="form-group">
-    {{ form.password.label() }}
-    {{ form.password() }}
-  </div>
-  <br />
-  {{ form.submit(class_="btn btn-primary") }}
+    <div class="page-header mb-3">
+        <h1>Log In</h1>
+    </div>
+    <div class="form-group mb-3">
+        {{ form.email.label() }}
+        {{ form.email() }}
+    </div>
+    <div class="form-group mb-3">
+        {{ form.password.label() }}
+        {{ form.password() }}
+    </div>
+    <br />
+    {{ form.submit(class_="btn btn-primary") }}
 </form>
 {% endblock %}
+
+{% block javascript %}
+<script>
+    let elements = document.getElementsByTagName("label");
+    for (let i = 0; i < elements.length; i++) {
+        elements[i].classList.add("form-label");
+    }
+</script>
+{% endblock javascript %}
