+++ website/templates/people.html
@@ -1,7 +1,7 @@
 {% extends "base.html" %}
 {% block title %}People{% endblock %} 
 {% block content %}
-  <form method="POST" class="container m-1">
+  <form method="POST" class="container">
     <div class="row align-items-center mb-4">
       <div class="col-12 col-md-4 col-xl-3">
         <h1>People</h1>
