+++ website/templates/sign-up.html
@@ -2,7 +2,7 @@
 {% block title %}Sign Up{% endblock %}
 {% block content %}
 <form method="POST">
-  <div class="page-header">
+  <div class="page-header mb-3">
     <h1>Sign Up</h1>
   </div>
   <div class="form-group">
